+++ CHANGELOG.md
@@ -22,6 +22,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mse304247: Attempting to use too long tag breaks popup](https://meta.stackexchange.com/q/304247)
 * [boardgames1652: Switch Magic autocard over to a different search engine, Scryfall](https://boardgames.meta.stackexchange.com/q/1652) (replaces [boardgames1152](https://boardgames.meta.stackexchange.com/q/1152))
 * [boardgames867: We should implement Magic the Gathering pop-ups on hover](https://boardgames.meta.stackexchange.com/q/867)
+* [mso360855: Margins are off/unbalanced in inbox](https://meta.stackoverflow.com/q/360855)
 
 **Removed fixes:**